requirements.txt
fastapi
uvicorn
minio
pymongo
python-dotenv